+++ tests/unit/test_main.py
@@ -8,11 +8,7 @@ from kytos.lib.helpers import (
     get_kytos_event_mock,
     get_switch_mock,
 )
-from napps.amlight.flow_stats.main import GenericFlow, Main
-from napps.kytos.of_core.v0x04.flow import Action as Action40
-from napps.kytos.of_core.v0x04.match_fields import MatchFieldFactory
-from napps.kytos.of_core.v0x04.flow import Match as Match40
-from pyof.foundation.basic_types import UBInt32
+from napps.amlight.flow_stats.main import Main
 
 
 # pylint: disable=too-many-public-methods, too-many-lines
@@ -24,7 +20,8 @@ class TestMain(TestCase):
 
         Set the server_name_url_url from amlight/flow_stats
         """
-        self.server_name_url = "http://localhost:8181/api/amlight/flow_stats"
+        self.server_name_url = \
+            "http://localhost:8181/api/amlight/flow_stats/v1"
         self.napp = Main(get_controller_mock())
 
     @staticmethod
@@ -59,210 +56,263 @@ class TestMain(TestCase):
             (
                 {"dpid": "[dpid]"},
                 {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/flow/match/ ",
-            ),
-            (
-                {"dpid": "[dpid]"},
-                {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/flow/stats/",
+                "/api/amlight/flow_stats/v1/flow/stats/",
             ),
             (
                 {"flow_id": "[flow_id]"},
                 {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/packet_count/",
+                "/api/amlight/flow_stats/v1/packet_count/",
             ),
             (
                 {"flow_id": "[flow_id]"},
                 {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/bytes_count/",
+                "/api/amlight/flow_stats/v1/bytes_count/",
             ),
             (
                 {"dpid": "[dpid]"},
                 {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/packet_count/per_flow/",
+                "/api/amlight/flow_stats/v1/packet_count/per_flow/",
             ),
             (
                 {"dpid": "[dpid]"},
                 {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/packet_count/sum/",
-            ),
-            (
-                {"dpid": "[dpid]"},
-                {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/bytes_count/per_flow/",
-            ),
-            (
-                {"dpid": "[dpid]"},
-                {"OPTIONS", "HEAD", "GET"},
-                "/api/amlight/flow_stats/bytes_count/sum/",
+                "/api/amlight/flow_stats/v1/bytes_count/per_flow/",
             ),
         ]
         urls = self.get_napp_urls(self.napp)
-        self.assertEqual(len(expected_urls), len(urls))
+        assert len(expected_urls) == len(urls)
+
+    def test_flow_from_id(self):
+        """Test flow_from_id function"""
+        flow = self._get_mocked_flow_base()
+        self.napp.flows_stats_dict = {
+            flow.id: flow
+        }
+        results = self.napp.flow_from_id(flow.id)
+        assert results.id == flow.id
+
+    def test_flow_from_id__fail(self):
+        """Test flow_from_id function"""
+        flow = self._get_mocked_flow_base()
+        self.napp.flows_stats_dict = {
+            flow.id: flow
+        }
+        results = self.napp.flow_from_id('1')
+        assert results is None
+
+    def test_flow_from_id__empty(self):
+        """Test flow_from_id function when flows_stats_dict is empty"""
+        self.napp.flows_stats_dict = {}
+        results = self.napp.flow_from_id('1')
+        assert results is None
 
     def test_packet_count__fail(self):
-        """Test packet_count rest call with wrong flow_id."""
+        """Test bytes_count rest call with wrong flow_id."""
         flow_id = "123456789"
         rest_name = "packet_count"
         response = self._get_rest_response(rest_name, flow_id)
 
-        self.assertEqual(response.data, b"Flow does not exist")
+        assert response.data == b"Flow does not exist"
 
-    def test_packet_count(self):
+    @patch("napps.amlight.flow_stats.main.Main.flow_from_id")
+    def test_packet_count(self, mock_from_flow):
         """Test packet_count rest call."""
-        flow_id = "1"
+        flow_id = '1'
+        dpid_id = '1'
+        mock_from_flow.return_value = self._get_mocked_flow_base()
+
         rest_name = "packet_count"
         self._patch_switch_flow(flow_id)
-        response = self._get_rest_response(rest_name, flow_id)
 
+        response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
-        self.assertEqual(json_response["flow_id"], flow_id)
-        self.assertEqual(json_response["packet_counter"], 40)
-        self.assertEqual(json_response["packet_per_second"], 2.0)
+        assert json_response["flow_id"] == flow_id
+        assert json_response["packet_counter"] == 40
+        assert json_response["packet_per_second"] == 2.0
 
-    def test_bytes_count_fail(self):
+    def test_bytes_count__fail(self):
         """Test bytes_count rest call with wrong flow_id."""
         flow_id = "123456789"
         rest_name = "bytes_count"
         response = self._get_rest_response(rest_name, flow_id)
 
-        self.assertEqual(response.data, b"Flow does not exist")
+        assert response.data == b"Flow does not exist"
 
-    def test_bytes_count(self):
+    @patch("napps.amlight.flow_stats.main.Main.flow_from_id")
+    def test_bytes_count(self, mock_from_flow):
         """Test bytes_count rest call."""
-        flow_id = "1"
+        flow_id = '1'
+        dpid_id = '1'
+        mock_from_flow.return_value = self._get_mocked_flow_base()
+
         rest_name = "bytes_count"
         self._patch_switch_flow(flow_id)
-        response = self._get_rest_response(rest_name, flow_id)
 
+        response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
-        self.assertEqual(json_response["flow_id"], flow_id)
-        self.assertEqual(json_response["bytes_counter"], 10)
-        self.assertEqual(json_response["bits_per_second"], 4.0)
+        assert json_response["flow_id"] == flow_id
+        assert json_response["bytes_counter"] == 10
+        assert json_response["bits_per_second"] == 4.0
 
-    def test_packet_count_per_flow(self):
+    def test_packet_count_per_flow__empty(self):
+        """Test packet_count rest call with a flow that does not exist ."""
+        flow_id = "123456789"
+        rest_name = "packet_count/per_flow"
+        response = self._get_rest_response(rest_name, flow_id)
+        json_response = json.loads(response.data)
+        assert len(json_response) == 0
+
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_packet_count_per_flow(self, mock_from_flow):
         """Test packet_count_per_flow rest call."""
-        flow_id = "1"
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
         rest_name = "packet_count/per_flow"
         self._patch_switch_flow(flow_id)
-
-        dpid_id = 111
-        response = self._get_rest_response(rest_name, dpid_id)
-
-        json_response = json.loads(response.data)
-        self.assertEqual(json_response[0]["flow_id"], flow_id)
-        self.assertEqual(json_response[0]["packet_counter"], 40)
-        self.assertEqual(json_response[0]["packet_per_second"], 2.0)
-
-    def test_packet_count_sum(self):
-        """Test packet_count_sum rest call."""
-        flow_id = "1"
-        rest_name = "packet_count/sum"
-        self._patch_switch_flow(flow_id)
-
-        dpid_id = 111
         response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
+        assert json_response[0]["flow_id"] == flow_id
+        assert json_response[0]["packet_counter"] == 40
+        assert json_response[0]["packet_per_second"] == 2.0
 
-        self.assertEqual(json_response, 40)
+    def test_bytes_count_per_flow__empty(self):
+        """Test bytes_count rest call with a flow that does not exist ."""
+        flow_id = "123456789"
+        rest_name = "bytes_count/per_flow"
+        response = self._get_rest_response(rest_name, flow_id)
+        json_response = json.loads(response.data)
+        assert len(json_response) == 0
 
-    def test_bytes_count_per_flow(self):
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_bytes_count_per_flow(self, mock_from_flow):
         """Test bytes_count_per_flow rest call."""
-        flow_id = "1"
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
+
         rest_name = "bytes_count/per_flow"
         self._patch_switch_flow(flow_id)
 
-        dpid_id = 111
-        response = self._get_rest_response(rest_name, dpid_id)
-
-        json_response = json.loads(response.data)
-        self.assertEqual(json_response[0]["flow_id"], flow_id)
-        self.assertEqual(json_response[0]["bytes_counter"], 10)
-        self.assertEqual(json_response[0]["bits_per_second"], 4.0)
-
-    def test_bytes_count_sum(self):
-        """Test bytes_count_sum rest call."""
-        flow_id = "1"
-        rest_name = "bytes_count/sum"
-        self._patch_switch_flow(flow_id)
-
-        dpid_id = 111
         response = self._get_rest_response(rest_name, dpid_id)
         json_response = json.loads(response.data)
+        assert json_response[0]["flow_id"] == flow_id
+        assert json_response[0]["bytes_counter"] == 10
+        assert json_response[0]["bits_per_second"] == 4.0
 
-        self.assertEqual(json_response, 10)
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_flows_counters_packet(self, mock_from_flow):
+        """Test flows_counters function for packet"""
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
 
-    @patch("napps.amlight.flow_stats.main.Main.match_flows")
-    def test_flow_match(self, mock_match_flows):
-        """Test flow_match rest call."""
-        flow = GenericFlow()
-        flow.actions = [
-            Action40.from_dict(
-                {
-                    "action_type": "output",
-                    "port": "1",
-                }
-            ),
-        ]
-        flow.version = "0x04"
-        mock_match_flows.return_value = flow
+        with get_controller_mock().api_server.app.app_context():
+            response = self.napp.flows_counters('packet_count', dpid_id)
+            json_response = response.json
+            assert len(json_response) == 1
 
-        flow_id = "1"
-        rest_name = "flow/match"
-        self._patch_switch_flow(flow_id)
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_flows_counters_bytes(self, mock_from_flow):
+        """Test flows_counters function for bytes"""
+        flow_stats = {
+            'byte_count': 10,
+            'duration_sec': 20,
+            'duration_nsec': 30,
+            'packet_count': 40
+            }
+        flow_id = '6055f13593fad45e0b4699f49d56b105'
+        flow_stats_dict_mock = {flow_id: flow_stats}
+        dpid_id = "00:00:00:00:00:00:00:01"
+        flow_by_sw = {dpid_id: flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
 
-        dpid_id = "aa:00:00:00:00:00:00:11"
-        response = self._get_rest_response(rest_name, dpid_id)
-        json_response = json.loads(response.data)
+        with get_controller_mock().api_server.app.app_context():
+            response = self.napp.flows_counters('byte_count', dpid_id)
+            json_response = response.json
+            assert len(json_response) == 1
 
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(json_response["actions"][0]["action_type"], "output")
-        self.assertEqual(json_response["actions"][0]["port"], "1")
-        self.assertEqual(json_response["version"], "0x04")
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_flow_stats_by_dpid_flow_id(self, mock_from_flow):
+        """Test flow_stats rest call."""
+        flow_stats = {
+            'byte_count': 148,
+            'duration_sec': 1589,
+            'duration_nsec': 556000000,
+            'packet_count': 2
+            }
+        flow_stats_dict_mock = {'6055f13593fad45e0b4699f49d56b105': flow_stats}
+        flow_by_sw = {"00:00:00:00:00:00:00:01": flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
 
-    @patch("napps.amlight.flow_stats.main.Main.match_flows")
-    def test_flow_match_fail(self, mock_match_flows):
-        """Test flow_match rest call."""
-        mock_match_flows.return_value = None
+        api = get_test_client(self.napp.controller, self.napp)
+        endpoint = "/flow/stats?dpid=00:00:00:00:00:00:00:01"
+        url = f"{self.server_name_url}"+endpoint
 
-        flow_id = "1"
-        rest_name = "flow/match"
-        self._patch_switch_flow(flow_id)
+        response = api.get(url)
+        expected = flow_by_sw
+        assert response.json == expected
+        assert response.status_code == 200
 
-        dpid_id = "aa:00:00:00:00:00:00:11"
-        response = self._get_rest_response(rest_name, dpid_id)
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_flow_stats_by_dpid_flow_id_with_dpid(self, mock_from_flow):
+        """Test flow_stats rest call."""
+        flow_stats = {
+            'byte_count': 148,
+            'duration_sec': 1589,
+            'duration_nsec': 556000000,
+            'packet_count': 2
+            }
+        flow_stats_dict_mock = {'6055f13593fad45e0b4699f49d56b105': flow_stats}
+        flow_by_sw = {"00:00:00:00:00:00:00:01": flow_stats_dict_mock}
+        mock_from_flow.return_value = flow_by_sw
 
-        self.assertEqual(response.status_code, 404)
+        api = get_test_client(self.napp.controller, self.napp)
+        endpoint = "/flow/stats?dpid=00:00:00:00:00:00:00:01"
+        url = f"{self.server_name_url}"+endpoint
 
-    @patch("napps.amlight.flow_stats.main.Main.match_flows")
-    def test_flow_stats(self, mock_match_flows):
-        """Test flow_match rest call."""
-        flow = GenericFlow()
-        flow.actions = [
-            Action40.from_dict(
-                {
-                    "action_type": "output",
-                    "port": "1",
-                }
-            ),
-        ]
-        flow.version = "0x04"
-        mock_match_flows.return_value = [flow]
+        response = api.get(url)
+        expected = flow_by_sw
+        assert response.json == expected
+        assert response.status_code == 200
 
-        flow_id = "1"
-        rest_name = "flow/stats"
-        self._patch_switch_flow(flow_id)
+    @patch("napps.amlight.flow_stats.main.Main.flow_stats_by_dpid_flow_id")
+    def test_flow_stats_by_dpid_flow_id_not_found(self, mock_from_flow):
+        """Test flow_stats rest call."""
+        flow_by_sw = {}
+        mock_from_flow.return_value = flow_by_sw
 
-        dpid_id = "aa:00:00:00:00:00:00:11"
-        response = self._get_rest_response(rest_name, dpid_id)
-        json_response = json.loads(response.data)
+        api = get_test_client(self.napp.controller, self.napp)
+        endpoint = "/flow/stats?dpid=00:00:00:00:00:00:00:01"
+        url = f"{self.server_name_url}"+endpoint
 
-        self.assertEqual(response.status_code, 200)
-        print(json_response)
-        self.assertEqual(json_response[0]["actions"][0]["action_type"],
-                         "output")
-        self.assertEqual(json_response[0]["actions"][0]["port"], "1")
-        self.assertEqual(json_response[0]["version"], "0x04")
+        response = api.get(url)
+        assert len(response.json) == 0
 
     def _patch_switch_flow(self, flow_id):
         """Helper method to patch controller to return switch/flow data."""
@@ -270,7 +320,6 @@ class TestMain(TestCase):
         flow = self._get_mocked_flow_stats()
         flow.id = flow_id
         switch = MagicMock()
-        switch.generic_flows = [flow]
         self.napp.controller.switches = {"1": switch}
         self.napp.controller.get_switch_by_dpid = MagicMock()
         self.napp.controller.get_switch_by_dpid.return_value = switch
@@ -318,33 +367,6 @@ class TestMain(TestCase):
         flow.stats = self._get_mocked_flow_stats()
         return flow
 
-    @patch("napps.amlight.flow_stats.main.GenericFlow.from_flow_stats")
-    def test_handle_stats_reply(self, mock_from_flow):
-        """Test handle_stats_reply rest call."""
-        mock_from_flow.return_value = self._get_mocked_flow_base()
-
-        def side_effect(event):
-            self.assertTrue(f"{event}", "amlight/flow_stats.flows_updated")
-            self.assertTrue(event.content["switch"], 111)
-
-        self.napp.controller = MagicMock()
-        self.napp.controller.buffers.app.put.side_effect = side_effect
-
-        msg = MagicMock()
-        msg.flags.value = 2
-        msg.body = [self._get_mocked_flow_stats()]
-        event_switch = MagicMock()
-        event_switch.generic_flows = []
-        event_switch.dpid = 111
-        self.napp.handle_stats_reply(msg, event_switch)
-
-        # Check if important trace dont trigger the event
-        # It means that the CP trace is the same to the DP trace
-        self.napp.controller.buffers.app.put.assert_called_once()
-
-        # Check mocked flow id
-        self.assertEqual(event_switch.generic_flows[0].id, 456)
-
     @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply_received")
     def test_handle_stats_received(self, mock_handle_stats):
         """Test handle_stats_received function."""
@@ -360,7 +382,7 @@ class TestMain(TestCase):
         mock_handle_stats.assert_called_once()
 
     @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply_received")
-    def test_handle_stats_received_fail(self, mock_handle_stats):
+    def test_handle_stats_received__fail(self, mock_handle_stats):
         """Test handle_stats_received function for
         fail when replies_flows is not in content."""
 
@@ -373,148 +395,10 @@ class TestMain(TestCase):
         self.napp.handle_stats_received(event)
         mock_handle_stats.assert_not_called()
 
-    @patch("napps.amlight.flow_stats.main.GenericFlow.from_replies_flows")
-    def test_handle_stats_reply_received(self, mock_from_flow):
+    def test_handle_stats_reply_received(self):
         """Test handle_stats_reply_received call."""
-        mock_from_flow.return_value = self._get_mocked_flow_base()
 
-        event_switch = MagicMock()
         flows_mock = self._get_mocked_multipart_replies_flows()
-        self.napp.handle_stats_reply_received(event_switch, flows_mock)
+        self.napp.handle_stats_reply_received(flows_mock)
 
-        self.assertEqual(event_switch.generic_flows[0].id, 456)
-
-
-# pylint: disable=too-many-public-methods, too-many-lines
-class TestGenericFlow(TestCase):
-    """Test the GenericFlow class."""
-
-    # pylint: disable=no-member
-    def test_from_flow_stats__x04(self):
-        """Test from_flow_stats method 0x04 version."""
-        flow_stats = MagicMock()
-
-        flow_stats.actions = [
-            Action40.from_dict(
-                {
-                    "action_type": "output",
-                    "port": UBInt32(1),
-                }
-            ).as_of_action(),
-        ]
-
-        result = GenericFlow.from_flow_stats(flow_stats)
-
-        self.assertEqual(result.idle_timeout, flow_stats.idle_timeout.value)
-        self.assertEqual(result.hard_timeout, flow_stats.hard_timeout.value)
-        self.assertEqual(result.priority, flow_stats.priority.value)
-        self.assertEqual(result.table_id, flow_stats.table_id.value)
-        self.assertEqual(result.duration_sec, flow_stats.duration_sec.value)
-        self.assertEqual(result.packet_count, flow_stats.packet_count.value)
-        self.assertEqual(result.byte_count, flow_stats.byte_count.value)
-        self.assertEqual(result.version, "0x04")
-
-    def test_from_replies_flows(self):
-        """Test from_replies_flows method 0x04 version."""
-        replies_flow = MagicMock()
-
-        action_dict = {
-            "action_type": "output",
-            "port": UBInt32(1),
-        }
-
-        actions = Action40.from_dict(action_dict).as_of_action()
-
-        instruction = MagicMock()
-        instruction.instruction_type = 'apply_actions'
-        instruction.actions = [actions]
-        replies_flow.instructions = [instruction]
-        match = Match40(42)
-        replies_flow.match = match
-        result = GenericFlow.from_replies_flows(replies_flow)
-
-        self.assertEqual(result.idle_timeout, replies_flow.idle_timeout)
-        self.assertEqual(result.hard_timeout, replies_flow.hard_timeout)
-        self.assertEqual(result.priority, replies_flow.priority)
-        self.assertEqual(result.table_id, replies_flow.table_id)
-        self.assertEqual(result.cookie, replies_flow.cookie)
-        self.assertEqual(result.duration_sec, replies_flow.stats.duration_sec)
-        self.assertEqual(result.packet_count, replies_flow.stats.packet_count)
-        self.assertEqual(result.byte_count, replies_flow.stats.byte_count)
-
-        # pylint: disable=too-many-public-methods, too-many-lines
-        match_expect = MatchFieldFactory.from_of_tlv(
-            match.as_of_match().oxm_match_fields[0]
-            )
-        self.assertEqual(result.actions[0], actions)
-        self.assertEqual(result.match["in_port"], match_expect)
-
-    def test_to_dict__x04(self):
-        """Test to_dict method 0x04 version."""
-        match = {}
-        match["in_port"] = MagicMock()
-        match["in_port"].value = 22
-
-        generic_flow = GenericFlow(
-            version="0x04",
-            match=match,
-        )
-
-        result = generic_flow.to_dict()
-        expected = {
-            "version": "0x04",
-            "in_port": 22,
-            "idle_timeout": 0,
-            "hard_timeout": 0,
-            "priority": 0,
-            "table_id": 255,
-            "cookie": None,
-            "buffer_id": None,
-            "actions": [],
-        }
-
-        self.assertEqual(result, expected)
-
-    def test_match_to_dict(self):
-        """Test match_to_dict method for 0x04 version."""
-        match = {}
-        match["in_port"] = MagicMock()
-        match["in_port"].value = 22
-        match["vlan_vid"] = MagicMock()
-        match["vlan_vid"].value = 123
-
-        generic_flow = GenericFlow(
-            version="0x04",
-            match=match,
-        )
-        result = generic_flow.match_to_dict()
-        expected = {"in_port": 22, "vlan_vid": 123}
-
-        self.assertEqual(result, expected)
-
-    def test_match_to_dict__empty_match(self):
-        """Test match_to_dict method for 0x04 version, with empty matches."""
-        generic_flow = GenericFlow(version="0x04")
-        result = generic_flow.match_to_dict()
-        self.assertEqual(result, {})
-
-    def test_id__x04(self):
-        """Test id method 0x04 version."""
-        match = {}
-        match["in_port"] = MagicMock()
-        match["in_port"].value = 22
-        match["vlan_vid"] = MagicMock()
-        match["vlan_vid"].value = 123
-
-        generic_flow = GenericFlow(
-            version="0x04",
-            match=match,
-            idle_timeout=1,
-            hard_timeout=2,
-            priority=6,
-            table_id=7,
-            cookie=8,
-            buffer_id=9,
-        )
-
-        self.assertEqual(generic_flow.id, "2d843f76b8b254fad6c6e2a114590440")
+        assert list(self.napp.flows_stats_dict.values())[0].id == 456